+++ connector/ce/utils/db-schema-ce/src/main/resources/db/migration-ce/V16.4.0__Release_16.4.0_CE.sql
@@ -1,0 +1,3 @@
+
+-- Migrate the existing transfer processes' callback address to avoid global transfer processes stalling
+update edc_data_plane set callback_address = 'http://localhost:11004/api/control';