+++ frontend/src/app/(authenticated)/contracts/[contractId]/transfer/components/initiate-transfer-form-schema.tsx
@@ -10,12 +10,14 @@
 import {initiateTransferCustomSchema} from '@/app/(authenticated)/contracts/[contractId]/transfer/components/initiate-transfer-custom-form';
 import {initiateTransferHttpSchema} from '@/app/(authenticated)/contracts/[contractId]/transfer/components/initiate-transfer-http-form';
 import {z} from 'zod';
+import {initiateTransferAzureStorageSchema} from './initiate-transfer-azure-storage-form';
 
-export type InitiateTransferType = 'HTTP' | 'CUSTOM_JSON';
+export type InitiateTransferType = 'HTTP' | 'CUSTOM_JSON' | 'AZURE_STORAGE';
 
 export const initiateTransferFormSchema = z.discriminatedUnion('transferType', [
   initiateTransferCustomSchema,
   initiateTransferHttpSchema,
+  initiateTransferAzureStorageSchema,
 ]);
 
 export type InitiateTransferFormValue = z.infer<